details: editor support for v4 format (up to 16383 objects)
Mirror the engine's v4 detail format in the SDK and make the editor export
adaptively: v3 (16-byte slots) when a level uses <=63 distinct detail objects,
v4 (20-byte slots, 14-bit ids) only when it exceeds that. A level "upgrades" to
v4 automatically, per export, and drops back to v3 if objects are removed.

- DetailFormat.h: dual-version layout (DetailSlot_v3 + wide DetailSlot,
  expand_v3/pack_v3, static_asserts), shared byte-for-byte with the engine.
- DetailManager.h: dm_max_objects -> 16383, object pool -> xr_vector.
- ESceneDOTools: .dm project bumped to DETMGR_VERSION 4 (reads v3 via expand_v3,
  saves wide); Export() branches v3/v4 on the distinct-used-object count.
- Widen the export/generate object-index plumbing from u8 to u16 (remap_object,
  GetObject/GetRandomObject, the color-group remap) so >255 objects are addressable.

Existing v3 .dm projects and exports are unchanged.

+++ Source/Editors/LevelEditor/Editor/Tools/Details/ESceneDOTools.h
@@ -69,8 +69,8 @@ class EDetailManager: public CDetailManager, public ESceneToolBase
 
     void        GetSlotRect(Frect& rect, int sx, int sz);
     void        GetSlotTCRect(Irect& rect, int sx, int sz);
-    u8          GetRandomObject(u32 color_index);
-    u8          GetObject(ColorIndexPairIt& CI, u8 id);
+    u16         GetRandomObject(u32 color_index);
+    u16         GetObject(ColorIndexPairIt& CI, u16 id);
 
     void        CalcClosestCount(int part, const Fcolor& C, SIndexDistVec& best);
     void        FindClosestIndex(const Fcolor& C, SIndexDistVec& best);
